adapters: extract shared LiveKit server base; dedupe JobContext + fixes
The SIP AgentServer and audio-stream AudioStreamServer were ~75% the same
file (inference bootstrap, load monitor, HTTP surface, FfiQueue lifecycle
loop, force-exit cleanup), and their JobContext dataclasses duplicated the
session wiring, observability tagging, and listener registry.

Pull the shared body into _server_base.py:
  - AgentServerBase owns the common server machinery; subclasses provide
    transport-specific hooks (_on_participant_connected, _dispatch_transport_event,
    _extra_routes/_worker_extra, _run, _start_call/_start_session) and the
    _active_map/_contexts/_ended_events/_ffi_global accessors.
  - JobContextBase owns the shared JobContext methods (session getter/setter,
    tagger, proc, set_metadata, on/_emit) with _transport_tag/_unit_label/logger
    class attrs and a _hangup_on_close() hook (SIP off-loop vs audio inline).

server.py 1236->531, audio_stream_server.py 1009->389, both back under 1k;
behavior preserved (full test suite green).

Also:
  - node _session_finalize.ts: make the close-wait timeout actually force-close
    instead of an unreachable catch fallback.
  - node audio_stream_context.ts: guard null session (parity with SIP) and drop
    a stray stderr debug write.
  - audio_stream/endpoint.rs: fix stale wait_for_playout_async doc.
