+++ .gitignore
@@ -56,3 +56,6 @@ crates/agent-transport-node/npm/*/agent-transport-*.tgz
 
 # Claude Code worktrees (created by `/worktree` workflows)
 .claude/worktrees/
+
+# uv lockfile (Python env is uv-managed locally; not checked in)
+crates/agent-transport-python/uv.lock
