+++ crates/agent-transport-python/adapters/agent_transport/sip/livekit/audio_stream_server.py
@@ -20,100 +20,36 @@ Configure Plivo XML to return:
         </Stream>
     </Response>
 
-CLI commands (matching LiveKit):
-    python agent.py start   — production mode (INFO logging)
-    python agent.py dev     — development mode (DEBUG for adapters/pipeline)
-    python agent.py debug   — full debug (including Rust transport)
-
-Shutdown behavior: see ``server.py`` for the shutdown model. Same
-``os._exit(0)`` + per-session flush requirement applies here.
+The shared server machinery (inference bootstrap, load monitor, HTTP
+surface, FfiQueue lifecycle loop, force-exit cleanup) lives in
+``_server_base.AgentServerBase``. This file carries only the
+audio-stream specifics: WebSocket endpoint creation and the
+fire-and-forget Plivo hangup. Shutdown model: see ``server.py``.
 """
 
 import asyncio
 import logging
 import os
-import signal
 import sys
-import threading
 import time
 from dataclasses import dataclass, field
 from typing import Any, Callable, Coroutine
 
 import prometheus_client
-from aiohttp import web
 
-from agent_transport import AudioStreamEndpoint, init_logging
+from agent_transport import AudioStreamEndpoint
 from agent_transport._event import FfiEvent
 from agent_transport._event_sink import _on_event_from_rust
 from agent_transport._ffi_queue import GLOBAL
-from livekit.agents.inference_runner import _InferenceRunner
-from livekit.agents.utils.hw import get_cpu_monitor
-from livekit.agents.utils import MovingAverage
-from ._audio_io import TransportAudioInput, TransportAudioOutput
 from ._room_facade import TransportJobContextMixin, TransportRoom, create_transport_context
-from ._aio_utils import call_setup as _call_setup
-from ._session_teardown import force_shutdown_agent_session
 from ._session_finalize import finalize_session
+from ._server_base import AgentServerBase, JobContextBase, _nodename
 from .judging import EvaluationConfig
-from livekit.rtc.room import SipDTMF
-from .server import JobProcess
-from .observability import _ensure_transport_tags, _get_observability_url
+from .observability import _get_observability_url
 
 logger = logging.getLogger("agent_transport.audio_stream_server")
 
-
-# ─── Shared helpers (reuse from server.py) ────────────────────────────────────
-
-_inference_ctx_token = None
-
-def _set_inference_context(executor) -> None:
-    global _inference_ctx_token
-    from livekit.agents.job import _JobContextVar
-
-    class _Stub:
-        @property
-        def inference_executor(self):
-            return executor
-
-    _inference_ctx_token = _JobContextVar.set(_Stub())
-
-
-def _clear_inference_context() -> None:
-    global _inference_ctx_token
-    if _inference_ctx_token is not None:
-        from livekit.agents.job import _JobContextVar
-        _JobContextVar.reset(_inference_ctx_token)
-        _inference_ctx_token = None
-
-
-def _create_inference_executor(loop: asyncio.AbstractEventLoop):
-    from livekit.agents.ipc.inference_proc_executor import InferenceProcExecutor
-    import multiprocessing as mp
-
-    runners = _InferenceRunner.registered_runners
-    if not runners:
-        return None
-
-    executor = InferenceProcExecutor(
-        runners=runners,
-        initialize_timeout=5 * 60,
-        close_timeout=5,
-        memory_warn_mb=2000,
-        memory_limit_mb=0,
-        ping_interval=5,
-        ping_timeout=60,
-        high_ping_threshold=2.5,
-        mp_ctx=mp.get_context("spawn"),
-        loop=loop,
-        http_proxy=None,
-    )
-    return executor
-
-
-# ─── Prometheus metrics ───────────────────────────────────────────────────────
-
-from livekit.agents.telemetry.metrics import RUNNING_JOB_GAUGE, CPU_LOAD_GAUGE
-from livekit.agents import utils as _lk_utils
+# ─── Audio-stream-specific Prometheus metrics ─────────────────────────────────
 
 STREAM_SESSIONS_TOTAL = prometheus_client.Counter(
     "lk_agents_audio_stream_sessions_total",
@@ -128,67 +64,26 @@ STREAM_SESSION_DURATION = prometheus_client.Histogram(
     buckets=[1, 5, 10, 30, 60, 120, 300, 600],
 )
 
-def _nodename() -> str:
-    return _lk_utils.nodename()
-
-
-def _get_sdk_version() -> str:
-    try:
-        from livekit.agents.version import __version__
-        return __version__
-    except ImportError:
-        return "unknown"
-
-
-class _LoadMonitor:
-    def __init__(self) -> None:
-        self._avg = MovingAverage(5)
-        self._cpu_monitor = get_cpu_monitor()
-        self._lock = threading.Lock()
-        self._stop = threading.Event()
-        self._thread = threading.Thread(target=self._sample_loop, daemon=True)
-        self._thread.start()
-
-    def _sample_loop(self) -> None:
-        # Cooperative shutdown via _stop event so the thread doesn't keep
-        # sampling CPU during process teardown. daemon=True still guarantees
-        # process exit, but a clean stop is friendlier to test harnesses.
-        while not self._stop.is_set():
-            cpu = self._cpu_monitor.cpu_percent(interval=0.5)
-            with self._lock:
-                self._avg.add_sample(cpu)
-
-    def get_load(self) -> float:
-        with self._lock:
-            return self._avg.get_avg()
-
-    def stop(self) -> None:
-        """Signal the sampler thread to exit and join briefly."""
-        self._stop.set()
-        self._thread.join(timeout=1.0)
-
-
-# ─── JobContext ───────────────────────────────────────────────────
 
 @dataclass
-class JobContext(TransportJobContextMixin):
+class JobContext(JobContextBase, TransportJobContextMixin):
     """Context passed to the @audio_stream_session handler.
 
-    Matches LiveKit's standard pattern exactly:
-        @server.audio_stream_session()
-        async def entrypoint(ctx: JobContext):
-            session = AgentSession(vad=..., stt=..., llm=..., tts=...)
-            ctx.session = session
-            await session.start(agent=Assistant(), room=ctx.room)
-
-    Setting ctx.session automatically wires audio stream I/O and registers
-    the close handler. Then session.start(room=ctx.room) works exactly
-    like LiveKit WebRTC.
+    Setting ctx.session automatically wires audio stream I/O and registers the
+    close handler; then ``session.start(room=ctx.room)`` works exactly like
+    LiveKit WebRTC. The shared method surface (session wiring, observability
+    tagging, listener registry) lives in :class:`JobContextBase`.
 
     DTMF events (equivalent of room.on("sip_dtmf_received") in WebRTC):
         job_ctx = get_job_context()
         job_ctx.room.on("sip_dtmf_received", handler)
     """
+
+    # JobContextBase hooks (plain class attrs — not dataclass fields).
+    _transport_tag = "audio_stream"
+    _unit_label = "Session"
+    _ctx_logger = logger
+    _debug_logger_name = "agent_transport.audio_stream"
 
     session_id: str
     plivo_call_uuid: str      # Plivo Call UUID
@@ -206,8 +101,8 @@ class JobContext(TransportJobContextMixin):
 
     _agent_name: str = field(default="agent", repr=False)
     # Stable developer-supplied id (typically UUID4). Set by the
-    # AudioStreamServer when creating the JobContext per session and
-    # threaded through to the observability emitter.
+    # AudioStreamServer when creating the JobContext per session and threaded
+    # through to the observability emitter.
     _agent_id: str = field(default="", repr=False)
     _session: Any = field(default=None, repr=False)
     _call_ended: asyncio.Event | None = field(default=None, repr=False)
@@ -216,123 +111,36 @@ class JobContext(TransportJobContextMixin):
     _event_listeners: dict = field(default_factory=dict, repr=False)
     # 0.2.x post-Tier-A: pointer to the process-global FfiQueue. The
     # constrained pyo3 dispatcher thread feeds it (one consumer of
-    # ``inner.events()``). Audio sources subscribe through it per-frame
-    # with a filter narrowing to (capture_audio_frame, source_handle).
+    # ``inner.events()``). Audio sources subscribe through it per-frame with a
+    # filter narrowing to (capture_audio_frame, source_handle).
     _events: Any = field(default=None, repr=False)
     _proc: Any = field(default=None, repr=False)
     _shutdown_callbacks: list = field(default_factory=list, repr=False)
 
-    @property
-    def session(self):
-        return self._session
-
-    @property
-    def tagger(self):
-        # ``_tagger`` is set on ``self`` by ``_init_transport_job_context``.
-        return getattr(self, "_tagger", None)
-
-    def set_metadata(self, metadata: dict[str, Any]) -> None:
-        """Attach session metadata to native LiveKit observability tags."""
-        self.metadata.update({str(key): value for key, value in metadata.items() if value is not None})
-        if account_id := self.metadata.get("account_id"):
-            self.account_id = str(account_id)
-            self.metadata["account_id"] = self.account_id
-
-        _ensure_transport_tags(
-            self.tagger,
-            account_id=self.account_id,
-            transport="audio_stream",
-            direction=self.direction,
-            agent_id=self._agent_id,
-            agent_name=self._agent_name,
-            metadata=self.metadata,
-        )
-
-    @session.setter
-    def session(self, session: Any) -> None:
-        """Set the agent session — automatically wires audio stream I/O.
-
-        This replaces the manual ctx.start() pattern. After setting ctx.session,
-        call session.start(agent=, room=ctx.room) directly.
-        """
-        self._session = session
-        self._primary_agent_session = session
-
-        # Wire audio stream I/O. TransportAudioOutput is a Pattern-A
-        # subclass of LiveKit's _ParticipantAudioOutput — buffering /
-        # forwarding / interrupt / playout logic are inherited verbatim,
-        # with our TransportAudioSource in place of rtc.AudioSource.
-        # ``events`` defaults to the process-global FfiQueue inside
-        # TransportAudioSource.
-        session.input.audio = TransportAudioInput(self.endpoint, self.session_id)
-        session.output.audio = TransportAudioOutput(
-            self.endpoint,
-            self.session_id,
-        )
-
-        # Listen to session close event — handles agent-initiated shutdown
-        @session.on("close")
-        def _on_session_close(ev):
-            logger.info("Session %s closed (reason=%s)", self.session_id, getattr(ev, 'reason', 'unknown'))
-            if self._call_ended is not None and not self._call_ended.is_set():
-                self._call_ended.set()
-            # AudioStreamEndpoint.hangup() is fire-and-forget in Rust: the Plivo
-            # REST DELETE is spawned on the endpoint's own tokio runtime and the
-            # call returns in microseconds. No Python thread (loop or executor)
-            # is held for the network round-trip, so calling it inline here does
-            # not stall the loop.
-            try:
-                self.endpoint.hangup(self.session_id)
-            except Exception:
-                pass
-
-        if logging.getLogger("agent_transport.audio_stream").isEnabledFor(logging.DEBUG):
-            @session.on("agent_state_changed")
-            def _on_agent_state(ev):
-                logger.info("Session %s agent: %s -> %s", self.session_id, ev.old_state, ev.new_state)
-            @session.on("user_state_changed")
-            def _on_user_state(ev):
-                logger.info("Session %s user: %s -> %s", self.session_id, ev.old_state, ev.new_state)
-
-    def on(self, event_name: str, callback: Callable | None = None) -> Callable:
-        """Register an event listener. Can be used as a decorator."""
-        def decorator(fn):
-            self._event_listeners.setdefault(event_name, []).append(fn)
-            return fn
-        if callback is not None:
-            return decorator(callback)
-        return decorator
-
-    def _emit(self, event_name: str, *args, **kwargs) -> None:
-        for listener in self._event_listeners.get(event_name, []):
-            try:
-                listener(*args, **kwargs)
-            except Exception:
-                logger.exception("Error in %s listener", event_name)
-
-    @property
-    def room(self):
-        """Room facade — use with session.start(room=ctx.room) like LiveKit WebRTC."""
-        return self._room
-
-    @property
-    def proc(self):
-        """Process context — access prewarm data via ctx.proc.userdata."""
-        return self._proc
-
-    def add_shutdown_callback(self, callback):
-        """Register a callback to run when the session ends."""
-        super().add_shutdown_callback(callback)
+    def _hangup_on_close(self) -> None:
+        # AudioStreamEndpoint.hangup() is fire-and-forget in Rust: the Plivo
+        # REST DELETE is spawned on the endpoint's own tokio runtime and the
+        # call returns in microseconds. No Python thread (loop or executor) is
+        # held for the network round-trip, so calling it inline is safe.
+        try:
+            self.endpoint.hangup(self.session_id)
+        except Exception:
+            pass
 
 
-# ─── AudioStreamServer ───────────────────────────────────────────────────────
-
-class AudioStreamServer:
+class AudioStreamServer(AgentServerBase):
     """Plivo audio streaming voice agent server.
 
     Equivalent of AgentServer but for Plivo WebSocket audio streaming.
     No SIP credentials needed — Plivo connects to your WebSocket server.
     """
+
+    _transport_name = "audio_stream"
+    _unit_label = "Session"
+    _worker_type = "JT_AUDIO_STREAM"
+    _transport_logger_name = "agent_transport.audio_stream"
+    _endpoint_not_ready_msg = "Audio stream endpoint not initialized"
+    _logger = logger
 
     def __init__(
         self,
@@ -343,10 +151,6 @@ class AudioStreamServer:
         sample_rate: int = 8000,
         host: str = "0.0.0.0",
         port: int | None = None,
-        # `agent_id` is the stable developer-supplied identifier — a UUID4
-        # (or any opaque string) that names this deployment uniquely across
-        # accounts. Mandatory: obs's agents view keys on it, and downstream
-        # `agent_transport_sessions.agent_id` is now NOT NULL.
         agent_id: str | None = None,
         agent_name: str = "audio-stream-agent",
         auth: Callable[..., bool | Coroutine] | None = None,
@@ -354,111 +158,44 @@ class AudioStreamServer:
         recording_dir: str = "/tmp/agent-sessions",
         recording_stereo: bool = True,
     ) -> None:
-        # Process-global FfiQueue (LiveKit-faithful mirror of
-        # ``FfiClient.instance.queue``). The pyo3 dispatcher thread
-        # spawned by ``set_event_sink`` does all the cross-thread
-        # ``call_soon_threadsafe`` plumbing — by the time events land
-        # here they're already on the asyncio loop.
-        self._events = GLOBAL
         self._listen_addr = listen_addr or os.environ.get("AUDIO_STREAM_ADDR", "0.0.0.0:8765")
         self._plivo_auth_id = plivo_auth_id or os.environ.get("PLIVO_AUTH_ID", "")
         self._plivo_auth_token = plivo_auth_token or os.environ.get("PLIVO_AUTH_TOKEN", "")
         self._sample_rate = sample_rate
-        self._host = host
-        self._port = port or int(os.environ.get("PORT", "8080"))
-        # Accept agent_id from explicit kwarg or AGENT_ID env var. We raise
-        # if neither is set rather than silently substituting a slug — the
-        # whole point of making this mandatory is to surface missing IDs
-        # loudly at server-start time instead of corrupting telemetry
-        # downstream.
-        resolved_agent_id = agent_id or os.environ.get("AGENT_ID") or None
-        if not resolved_agent_id:
-            raise ValueError(
-                "AudioStreamServer requires `agent_id` — pass a stable identifier "
-                "(typically a UUID4) via the `agent_id=` kwarg or the AGENT_ID env "
-                "var. This is the value that keys the obs agents view."
-            )
-        self._agent_id = resolved_agent_id
-        self._agent_name = agent_name
-        self._auth = auth
-        self._recording = recording
-        self._recording_dir = recording_dir
-        self._recording_stereo = recording_stereo
-        self._entrypoint_fnc: Callable[..., Coroutine] | None = None
-        self._setup_fnc: Callable | None = None
-        self._proc = JobProcess()
-        self._userdata: dict[str, Any] = {}
-        self._ep: AudioStreamEndpoint | None = None
-        # Session IDs are strings (Rust CallSession.session_id). Type hints
-        # were `int` previously — duck-typed at runtime, fixing the
-        # annotations so mypy/pyright agree with reality.
         self._active_sessions: dict[str, asyncio.Task] = {}
         self._session_ended_events: dict[str, asyncio.Event] = {}
         self._session_contexts: dict[str, JobContext] = {}
-        # Strong-reference set for fire-and-forget asyncio tasks (session
-        # start dispatch). Python's event loop only holds weak references
-        # to tasks; without storing them here, the GC can collect a task
-        # mid-execution and emit "Task was destroyed but it is pending!"
-        # warnings.
-        self._background_tasks: set[asyncio.Task] = set()
-        self._load_monitor = _LoadMonitor()
-        # Server-level event listeners. Plivo's audio_stream protocol has
-        # no pre-answer phase, so `ringing` doesn't fire here — but we
-        # expose the same `on(event_name)` shape as AgentServer for API
-        # symmetry. Future events (e.g., "session_start") can be added.
-        self._server_listeners: dict[str, list[Callable]] = {}
+        self._init_common(
+            host=host,
+            port=port,
+            agent_id=agent_id,
+            agent_name=agent_name,
+            auth=auth,
+            recording=recording,
+            recording_dir=recording_dir,
+            recording_stereo=recording_stereo,
+            events=GLOBAL,
+        )
+
+    # ── Base hooks ──
 
     @property
-    def setup_fnc(self):
-        return self._setup_fnc
+    def _active_map(self) -> dict:
+        return self._active_sessions
 
-    @setup_fnc.setter
-    def setup_fnc(self, fn):
-        """Set prewarm function — fn(proc: JobProcess). Matches LiveKit's server.setup_fnc = prewarm."""
-        self._setup_fnc = fn
+    @property
+    def _contexts(self) -> dict:
+        return self._session_contexts
 
-    def setup(self) -> Callable:
-        """Decorator to register a setup function that runs once at startup.
+    @property
+    def _ended_events(self) -> dict:
+        return self._session_ended_events
 
-        Example::
-            @server.setup()
-            def prewarm():
-                return {"vad": silero.VAD.load(), "turn_detector": MultilingualModel()}
-        """
-        def decorator(fn: Callable) -> Callable:
-            self._setup_fnc = fn
-            return fn
-        return decorator
+    def _ffi_global(self):
+        return GLOBAL
 
-    def on(self, event_name: str, callback: Callable | None = None) -> Callable:
-        """Register a server-level event listener.
-
-        Mirrors :meth:`AgentServer.on` for API symmetry. Plivo's
-        audio_stream protocol doesn't surface a pre-answer ringing phase
-        — Plivo only opens the WebSocket after the PSTN call is already
-        up — so there's no ``"ringing"`` event on this transport. The
-        hook shape exists for forward compatibility and so user code can
-        share handlers between the two server types.
-        """
-        def decorator(fn: Callable) -> Callable:
-            self._server_listeners.setdefault(event_name, []).append(fn)
-            return fn
-        if callback is not None:
-            return decorator(callback)
-        return decorator
-
-    def _emit_server_event(self, event_name: str, *args, **kwargs) -> None:
-        """Fire a server-level event to all registered listeners."""
-        listeners = self._server_listeners.get(event_name, [])
-        for listener in listeners:
-            try:
-                result = listener(*args, **kwargs)
-                if asyncio.iscoroutine(result):
-                    task = asyncio.create_task(result)
-                    self._background_tasks.add(task)
-                    task.add_done_callback(self._background_tasks.discard)
-            except Exception:
-                logger.exception("Server event listener for %r failed", event_name)
+    def _worker_extra(self) -> dict:
+        return {"listen_addr": self._listen_addr}
 
     def audio_stream_session(self) -> Callable:
         """Decorator to register the session handler."""
@@ -467,60 +204,37 @@ class AudioStreamServer:
             return fn
         return decorator
 
-    def run(self, port: int | None = None) -> None:
-        """Build CLI and run."""
-        if port is not None:
-            self._port = port
+    def on(self, event_name: str, callback: Callable | None = None) -> Callable:
+        """Register a server-level event listener.
 
-        try:
-            import typer
-            from typing import Annotated
-        except ImportError:
-            asyncio.run(self._run(log_mode="start"))
-            return
+        Mirrors :meth:`AgentServer.on` for API symmetry. Plivo's audio_stream
+        protocol has no pre-answer phase (Plivo only opens the WebSocket after
+        the PSTN call is already up), so there's no ``"ringing"`` event here —
+        the hook shape exists for forward compatibility and so user code can
+        share handlers between the two server types.
+        """
+        return super().on(event_name, callback)
 
-        app = typer.Typer()
+    def _on_participant_connected(self, e: FfiEvent, room_handle: str) -> bool:
+        info = e.room_event.participant_connected.info
+        session_id = info.session_id or room_handle
+        plivo_call_uuid = info.identity
+        stream_id = info.stream_id
+        extra_headers = info.extra_headers or {}
+        if session_id in self._active_sessions:
+            return False  # duplicate event or retry
+        logger.info(
+            "Audio stream session %s connected (plivo_call_uuid=%s, stream_id=%s)",
+            session_id, plivo_call_uuid, stream_id,
+        )
+        t = asyncio.create_task(
+            self._start_session(session_id, plivo_call_uuid, stream_id, extra_headers)
+        )
+        self._background_tasks.add(t)
+        t.add_done_callback(self._background_tasks.discard)
+        return False
 
-        @app.command()
-        def start(
-            port: Annotated[int | None, typer.Option(help="HTTP server port", envvar="PORT")] = None,
-        ) -> None:
-            """Run in production mode (INFO logging)."""
-            if port is not None:
-                self._port = port
-            asyncio.run(self._run(log_mode="start"))
-
-        @app.command()
-        def dev(
-            port: Annotated[int | None, typer.Option(help="HTTP server port", envvar="PORT")] = None,
-        ) -> None:
-            """Run in development mode (DEBUG for adapters/pipeline, INFO for Rust)."""
-            if port is not None:
-                self._port = port
-            asyncio.run(self._run(log_mode="dev"))
-
-        @app.command()
-        def debug(
-            port: Annotated[int | None, typer.Option(help="HTTP server port", envvar="PORT")] = None,
-        ) -> None:
-            """Run in debug mode (DEBUG everything including Rust transport)."""
-            if port is not None:
-                self._port = port
-            asyncio.run(self._run(log_mode="debug"))
-
-        @app.command(name="download-files")
-        def download_files() -> None:
-            """Download model files for plugins (turn detection, VAD, etc.)."""
-            import logging as _logging
-            from livekit.agents import Plugin
-
-            _logging.basicConfig(level=_logging.DEBUG)
-            for plugin in Plugin.registered_plugins:
-                logger.info("Downloading files for %s", plugin.package)
-                plugin.download_files()
-                logger.info("Finished downloading files for %s", plugin.package)
-
-        app()
+    # ── Audio-stream-specific run / orchestration ──
 
     async def _run(self, *, log_mode: str = "start") -> None:
         self._configure_logging(log_mode)
@@ -537,30 +251,10 @@ class AudioStreamServer:
 
         loop = asyncio.get_running_loop()
 
-        # Initialize inference executor
-        self._inference_executor = _create_inference_executor(loop)
-        if self._inference_executor:
-            await self._inference_executor.start()
-            await self._inference_executor.initialize()
-            logger.info("Inference executor ready (turn detection models available)")
+        # Inference executor + user prewarm (shared bootstrap).
+        await self._bootstrap_inference_and_setup(loop)
 
-        # Run user's setup function (supports sync and async)
-        if self._setup_fnc:
-            if self._inference_executor:
-                _set_inference_context(self._inference_executor)
-            try:
-                await _call_setup(self._setup_fnc, self._proc)
-            except Exception:
-                logger.exception("Setup function failed")
-                if self._inference_executor:
-                    _clear_inference_context()
-                raise
-            if self._inference_executor:
-                _clear_inference_context()
-            self._userdata = self._proc.userdata
-            logger.info("Setup complete: %s", list(self._userdata.keys()))
-
-        # Create AudioStreamEndpoint (starts WS server immediately)
+        # Create AudioStreamEndpoint (starts WS server immediately).
         self._ep = AudioStreamEndpoint(
             listen_addr=self._listen_addr,
             plivo_auth_id=self._plivo_auth_id,
@@ -576,323 +270,22 @@ class AudioStreamServer:
 
         # Wire the constrained pyo3 sink — spawns a dispatcher thread inside
         # Rust that drains ``inner.events()``, translates each event to a
-        # LiveKit-shape :class:`FfiEvent` and ``GLOBAL.put``s it on the
-        # asyncio loop via ``call_soon_threadsafe``. Replaces the legacy
-        # ``run_in_executor(wait_for_event)`` pump, cutting 2-3 asyncio
-        # loop ticks per event.
+        # LiveKit-shape :class:`FfiEvent` and ``GLOBAL.put``s it on the asyncio
+        # loop via ``call_soon_threadsafe``. Replaces the legacy
+        # ``run_in_executor(wait_for_event)`` pump, cutting 2-3 asyncio loop
+        # ticks per event.
         self._ep.set_event_sink(_on_event_from_rust)
 
-        # Start HTTP server
-        http_app = self._build_http_app()
-        runner = web.AppRunner(http_app)
-        await runner.setup()
-        site = web.TCPSite(runner, self._host, self._port, reuse_address=True)
-        await site.start()
-        logger.info("HTTP server on http://%s:%d", self._host, self._port)
-
-        # Start lifecycle loop (subscribes to GLOBAL FfiQueue).
-        event_task = asyncio.create_task(self._lifecycle_loop())
-
-        # Wait for shutdown signal
-        stop = asyncio.Event()
-        for sig in (signal.SIGINT, signal.SIGTERM):
-            loop.add_signal_handler(sig, stop.set)
-
-        await stop.wait()
-        try:
-            await self._run_cleanup(runner, event_task, loop)
-        finally:
-            # Flush stdio so the last log lines aren't lost —
-            # os._exit skips normal Python finalization.
-            try:
-                sys.stdout.flush()
-                sys.stderr.flush()
-            except Exception:
-                pass
-            os._exit(0)
-
-    async def _run_cleanup(
-        self,
-        runner: "web.AppRunner",
-        event_task: asyncio.Task,
-        loop: asyncio.AbstractEventLoop,
-    ) -> None:
-        """Hang up active sessions, drain ancillary resources with short timeouts.
-
-        Split out of ``run()`` so the signal-handler path remains a thin
-        wrapper that adds ``os._exit(0)`` after this returns, while tests can
-        exercise the cleanup ordering directly.
-        """
-        logger.info("Shutting down...")
-        if self._ep is not None:
-            # Hang up every active session up-front so callers drop promptly
-            # before the slower cleanup steps below. Unlike SIP, audio_stream
-            # hangup is fire-and-forget in Rust (returns instantly; the REST
-            # DELETE is flushed by ep.shutdown()'s drain window), so this inline
-            # loop never blocks the loop thread. ep.shutdown() below repeats it
-            # idempotently.
-            for session_id in list(self._active_sessions.keys()):
-                try:
-                    self._ep.hangup(session_id)
-                except Exception:
-                    pass
-        event_task.cancel()
-        try:
-            await asyncio.wait_for(runner.cleanup(), timeout=2.0)
-        except Exception:
-            pass
-        if self._inference_executor:
-            try:
-                await asyncio.wait_for(self._inference_executor.aclose(), timeout=2.0)
-            except Exception:
-                pass
-        self._load_monitor.stop()
-        if self._ep is not None:
-            try:
-                await asyncio.wait_for(
-                    loop.run_in_executor(None, self._ep.shutdown),
-                    timeout=2.0,
-                )
-            except Exception:
-                pass
-
-    def _configure_logging(self, mode: str) -> None:
-        if mode == "debug":
-            logging.basicConfig(
-                level=logging.DEBUG,
-                format="%(asctime)s.%(msecs)03d %(levelname)s %(name)s %(message)s",
-                datefmt="%H:%M:%S",
-                force=True,
-            )
-            init_logging(os.environ.get("RUST_LOG", "debug"))
-        elif mode == "dev":
-            logging.basicConfig(
-                level=logging.INFO,
-                format="%(asctime)s.%(msecs)03d %(levelname)s %(name)s %(message)s",
-                datefmt="%H:%M:%S",
-                force=True,
-            )
-            logging.getLogger("agent_transport.audio_stream").setLevel(logging.DEBUG)
-            logging.getLogger("livekit.agents").setLevel(logging.DEBUG)
-            logging.getLogger("livekit.plugins").setLevel(logging.DEBUG)
-            init_logging(os.environ.get("RUST_LOG", "info"))
-        else:
-            logging.basicConfig(
-                level=logging.INFO,
-                format="%(asctime)s %(levelname)s %(name)s %(message)s",
-                force=True,
-            )
-            init_logging(os.environ.get("RUST_LOG", "info"))
-
-    def _build_http_app(self) -> web.Application:
-        app = web.Application()
-        app.add_routes([
-            web.get("/", self._health_handler),
-            web.get("/worker", self._worker_handler),
-            web.get("/metrics", self._metrics_handler),
-        ])
-        return app
-
-    async def _check_auth(self, request: web.Request) -> web.Response | None:
-        if self._auth is None:
-            return None
-        result = self._auth(request)
-        if asyncio.iscoroutine(result):
-            result = await result
-        if result:
-            return None
-        return web.json_response({"error": "unauthorized"}, status=401)
-
-    async def _metrics_handler(self, request: web.Request) -> web.Response:
-        if err := await self._check_auth(request):
-            return err
-        loop = asyncio.get_running_loop()
-        node = _nodename()
-        CPU_LOAD_GAUGE.labels(nodename=node).set(self._load_monitor.get_load())
-        RUNNING_JOB_GAUGE.labels(nodename=node).set(len(self._active_sessions))
-
-        data = await loop.run_in_executor(None, prometheus_client.generate_latest)
-        return web.Response(
-            body=data,
-            headers={
-                "Content-Type": prometheus_client.CONTENT_TYPE_LATEST,
-                "Content-Length": str(len(data)),
-            },
-        )
-
-    async def _health_handler(self, request: web.Request) -> web.Response:
-        if not self._ep:
-            return web.Response(status=503, text="Audio stream endpoint not initialized")
-        return web.Response(text="OK")
-
-    async def _worker_handler(self, request: web.Request) -> web.Response:
-        if err := await self._check_auth(request):
-            return err
-        return web.json_response({
-            "agent_name": self._agent_name,
-            "worker_type": "JT_AUDIO_STREAM",
-            "worker_load": self._load_monitor.get_load(),
-            "active_jobs": len(self._active_sessions),
-            "sdk_version": _get_sdk_version(),
-            "project_type": "python",
-            "listen_addr": self._listen_addr,
-        })
-
-    async def _lifecycle_loop(self) -> None:
-        """Subscribe to GLOBAL FfiQueue for lifecycle events.
-
-        The pyo3 dispatcher thread (started by ``set_event_sink``) drains
-        ``inner.events()`` on the Rust side, translates each event to a
-        :class:`FfiEvent`, and ``GLOBAL.put``s on the asyncio loop via
-        ``call_soon_threadsafe``. By the time we ``await q.get()`` the
-        event is already on the loop — no ``run_in_executor`` round-trip,
-        so 2-3 asyncio ticks are saved per event vs. the previous
-        wait_for_event-based pump.
-        """
-        loop = asyncio.get_running_loop()
-
-        def _filter(e: FfiEvent) -> bool:
-            kind = e.WhichOneof("message")
-            if kind == "room_event":
-                return True
-            if kind == "transport_event":
-                sub = e.transport_event.WhichOneof("event")
-                return sub in (
-                    "beep_detected", "beep_timeout",
-                    "endpoint_shutdown", "call_ringing",
-                )
-            return False
-
-        q = GLOBAL.subscribe(loop=loop, filter_fn=_filter)
-        try:
-            while True:
-                e = await q.get()
-                try:
-                    if self._dispatch_event(e):
-                        break
-                except Exception:
-                    logger.exception("Error handling audio_stream FfiEvent %r", e.WhichOneof("message") if e else e)
-                finally:
-                    q.task_done()
-        except asyncio.CancelledError:
-            pass
-        finally:
-            GLOBAL.unsubscribe(q)
-
-    def _dispatch_event(self, e: FfiEvent) -> bool:
-        """Process a single FfiEvent. Returns True if the loop should exit."""
-        kind = e.WhichOneof("message")
-
-        if kind == "transport_event":
-            sub = e.transport_event.WhichOneof("event")
-            if sub == "endpoint_shutdown":
-                logger.debug("audio_stream lifecycle loop received endpoint_shutdown")
-                return True
-            if sub == "beep_detected":
-                be = e.transport_event.beep_detected
-                session_id = be.source_handle
-                logger.info(
-                    "Beep detected on session %s (freq=%.0fHz, dur=%dms)",
-                    session_id, be.frequency_hz, be.duration_ms,
-                )
-                ctx = self._session_contexts.get(session_id)
-                if ctx:
-                    ctx._emit("beep_detected", be.frequency_hz, be.duration_ms)
-                    if ctx._room:
-                        ctx._room.emit("beep_detected", {
-                            "frequency_hz": be.frequency_hz,
-                            "duration_ms": be.duration_ms,
-                        })
-                return False
-            if sub == "beep_timeout":
-                session_id = e.transport_event.beep_timeout.source_handle
-                logger.debug("Beep timeout on session %s", session_id)
-                ctx = self._session_contexts.get(session_id)
-                if ctx:
-                    ctx._emit("beep_timeout")
-                    if ctx._room:
-                        ctx._room.emit("beep_timeout", {})
-                return False
-            return False
-
-        if kind == "room_event":
-            sub = e.room_event.WhichOneof("participant")
-            room_handle = e.room_event.room_handle
-
-            if sub == "participant_connected":
-                info = e.room_event.participant_connected.info
-                session_id = info.session_id or room_handle
-                plivo_call_uuid = info.identity
-                stream_id = info.stream_id
-                extra_headers = info.extra_headers or {}
-                if session_id in self._active_sessions:
-                    return False  # duplicate event or retry
-                logger.info(
-                    "Audio stream session %s connected (plivo_call_uuid=%s, stream_id=%s)",
-                    session_id, plivo_call_uuid, stream_id,
-                )
-                t = asyncio.create_task(
-                    self._start_session(session_id, plivo_call_uuid, stream_id, extra_headers)
-                )
-                self._background_tasks.add(t)
-                t.add_done_callback(self._background_tasks.discard)
-                return False
-
-            if sub == "participant_disconnected":
-                pd = e.room_event.participant_disconnected
-                session_id = pd.session_id or room_handle
-                logger.info("Session %s terminated (reason=%s)", session_id, pd.reason)
-                try:
-                    self._ep.clear_buffer(session_id)
-                except Exception:
-                    pass
-                # Synchronously begin tearing down the AgentSession so a
-                # buffered STT transcript delivered after disconnect can't
-                # trigger a wasted LLM + TTS turn on a dead session (#83).
-                # Must run here, on the wake branch, before we set the
-                # session-ended event — flipping the scheduling guard later
-                # (in _run_session's finally) would be too late.
-                ctx = self._session_contexts.get(session_id)
-                if ctx is not None and getattr(ctx, "_session", None) is not None:
-                    force_shutdown_agent_session(ctx._session, self._background_tasks)
-
-                # Wake _run_session (which holds _JobContextVar). The
-                # ``participant_disconnected`` emit must come from
-                # _run_session — not here — because LiveKit's RoomIO
-                # synchronously calls ``AgentSession._close_soon`` which
-                # captures the current task's context.
-                if session_id in self._session_ended_events:
-                    self._session_ended_events[session_id].set()
-                return False
-
-            if sub == "data_packet_received":
-                dp = e.room_event.data_packet_received.value
-                if dp and dp.sip_dtmf:
-                    session_id = room_handle
-                    digit = dp.sip_dtmf.digit
-                    logger.debug("DTMF '%s' on session %s", digit, session_id)
-                    ctx = self._session_contexts.get(session_id)
-                    if ctx:
-                        ctx._emit("dtmf_received", digit)
-                        if ctx._room:
-                            dtmf_ev = SipDTMF(
-                                code=ord(digit) if digit else 0, digit=digit,
-                                participant=ctx._room._remote,
-                            )
-                            ctx._room.emit("sip_dtmf_received", dtmf_ev)
-                return False
-            return False
-
-        return False
+        await self._serve_until_shutdown(loop)
 
     async def _start_session(self, session_id: str, plivo_call_uuid: str, stream_id: str, extra_headers: dict) -> None:
         session_ended = asyncio.Event()
         self._session_ended_events[session_id] = session_ended
 
-        # Create Room facade BEFORE handler runs — ctx.room is available immediately.
-        # remote_kind=0 (STANDARD) because Plivo audio_stream is a WebSocket
-        # transport, not SIP — `participant.kind` should reflect that for any
-        # agent code that inspects it.
+        # Create Room facade BEFORE handler runs — ctx.room is available
+        # immediately. remote_kind=0 (STANDARD) because Plivo audio_stream is a
+        # WebSocket transport, not SIP — `participant.kind` should reflect that
+        # for any agent code that inspects it.
         room = TransportRoom(
             self._ep, session_id,
             agent_name=self._agent_name,
@@ -927,12 +320,10 @@ class AudioStreamServer:
 
         async def _run_session():
             # Re-set _JobContextVar in this task's own context so late
-            # ``session.close`` listeners (e.g.
-            # ``livekit.agents.beta.tools.end_call._on_session_close``)
-            # find the context regardless of how the close emit is
-            # scheduled. The parent context's set() returns a token
-            # scoped to the parent — child tasks inherit the value but
-            # not always reliably under heavy async churn.
+            # ``session.close`` listeners find the context regardless of how the
+            # close emit is scheduled. The parent context's set() returns a
+            # token scoped to the parent — child tasks inherit the value but not
+            # always reliably under heavy async churn.
             from livekit.agents.job import _JobContextVar
             _JobContextVar.set(ctx)
 
@@ -940,7 +331,7 @@ class AudioStreamServer:
             STREAM_SESSIONS_TOTAL.labels(nodename=node).inc()
             session_start = time.monotonic()
 
-            # Start recording if enabled
+            # Start recording if enabled.
             rec_path = None
             rec_started_at = None
             if self._recording:
@@ -955,8 +346,8 @@ class AudioStreamServer:
 
             try:
                 await self._entrypoint_fnc(ctx)
-                # Entrypoint returned — session.start() is non-blocking,
-                # so wait for stream to actually end (Plivo stop or agent shutdown)
+                # Entrypoint returned — session.start() is non-blocking, so wait
+                # for the stream to actually end (Plivo stop or agent shutdown).
                 if session_ended and not session_ended.is_set():
                     await session_ended.wait()
             except Exception:
@@ -984,22 +375,11 @@ class AudioStreamServer:
                     pass
                 # Cleanup Room facade. We do NOT call
                 # ``_JobContextVar.reset(job_ctx_token)`` here: late
-                # ``session.close`` listeners (e.g.
-                # ``livekit.agents.beta.tools.end_call._on_session_close``)
-                # can fire AFTER ``session.aclose()`` returns and they
-                # need ``get_job_context()`` to work. The contextvar is
-                # scoped to this asyncio.Task and dies cleanly when the
-                # task exits, so explicit reset is unnecessary.
+                # ``session.close`` listeners can fire AFTER ``session.aclose()``
+                # returns and they need ``get_job_context()`` to work. The
+                # contextvar is scoped to this asyncio.Task and dies cleanly
+                # when the task exits.
                 room._on_session_ended()
-                # No EventWaiter to close in the LiveKit-faithful model —
-                # any in-flight ``capture_frame`` / ``wait_for_playout``
-                # is awaiting on its own per-call subscribed Queue. The
-                # Rust ``AudioBuffer::Drop`` emits an
-                # ``audio_capture_error { error: "buffer_dropped" }`` for
-                # every pending async_id when the session is torn down,
-                # which is dispatched through the FfiQueue and resolves
-                # each awaiter with a ``RuntimeError``. Per-call
-                # unsubscribe runs in the audio source's ``finally``.
                 self._active_sessions.pop(session_id, None)
                 self._session_ended_events.pop(session_id, None)
                 self._session_contexts.pop(session_id, None)
